# users/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de usúarios",
        "description": "Esta API foi criada para o teste prático de um processo seletivo, esta API permite listar e cadastrar usúarios.",
        "contact": {
            "email": "jg.scheleder@hotmail.com"
        }
    },
    "basePath": "/v1",
    "tags": [{
        "name": "user" ,
        "description" : "Manipula os usúarios do sistema" }],
    "servers": [{
            "url": "http://localhost:8080",
            "description": "URL de testes"
            },
            {
                "url": "https://cargon.com.br",
                "description": "URL de produção"
            }
        ],
    
    "paths": {
        "/v1/users": {
            "get" : {
                "summary": "Lista todos os usúarios",
                "consumes" : "application/json",
                "produces": "application/json",
                "tags": ["user"],
                "responses": {
                    "200": {
                      "description": "OK",
                      "content": {
                        "application/json": {
                            "schema" : {
                                "type": "array",
                                "items" : {
                                    "$ref" : "#/components/schemas/userResponse"
                                }
                        }
                      }
                    }
                },
                    "500": {
                        "description": "Internal server error"}
                }
            }       
        },
        "/v1/user/create": {
            "post" : {
                "summary": "Cadastra um usúario",
                "consumes" : "application/json",
                "produces": "application/json",
                "tags": ["user"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema" : {
                                "$ref" : "#/components/schemas/user"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                      "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"},
                    "409": {
                        "description": "Conflict, resource already exists"},
                    "500": {
                        "description": "Internal server error"}

                }
            }       
        }       
    }
,
    "components": {
        "schemas": {
            "user": {
                "type":"object",
                "properties": {
                    "name" :{"type": "string"},
                    "email" :{"type": "string"},
                    "password": {"type":"string"}

                }
            },
            "userResponse": {
                "type":"object",
                "properties": {
                    "id" :{"type": "number"},
                    "name" :{"type": "string"},
                    "email" :{"type": "string"}

                }
            }
        }
    }
}